requirements.txt
streamlit
easyocr
Pillow
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
sentence-transformers
python-decouple
faiss-cpu
protobuf==3.20.*
